requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==1.10.13